+++ user_interface.ui
@@ -10,19 +10,19 @@
     <x>0</x>
     <y>0</y>
     <width>800</width>
-    <height>600</height>
+    <height>700</height>
    </rect>
   </property>
   <property name="minimumSize">
    <size>
     <width>800</width>
-    <height>600</height>
+    <height>700</height>
    </size>
   </property>
   <property name="maximumSize">
    <size>
     <width>800</width>
-    <height>600</height>
+    <height>700</height>
    </size>
   </property>
   <property name="cursor">
@@ -35,8 +35,8 @@
    <widget class="QPushButton" name="launch">
     <property name="geometry">
      <rect>
-      <x>150</x>
-      <y>150</y>
+      <x>10</x>
+      <y>69</y>
       <width>75</width>
       <height>23</height>
      </rect>
@@ -51,8 +51,8 @@
    <widget class="QLabel" name="warning_text_launch">
     <property name="geometry">
      <rect>
-      <x>250</x>
-      <y>151</y>
+      <x>110</x>
+      <y>70</y>
       <width>161</width>
       <height>20</height>
      </rect>
@@ -71,8 +71,8 @@
    <widget class="QSpinBox" name="spinBox_white_point">
     <property name="geometry">
      <rect>
-      <x>521</x>
-      <y>160</y>
+      <x>531</x>
+      <y>110</y>
       <width>81</width>
       <height>22</height>
      </rect>
@@ -90,8 +90,8 @@
    <widget class="QLabel" name="label_4">
     <property name="geometry">
      <rect>
-      <x>520</x>
-      <y>140</y>
+      <x>530</x>
+      <y>90</y>
       <width>61</width>
       <height>16</height>
      </rect>
@@ -103,8 +103,8 @@
    <widget class="QPushButton" name="select_file">
     <property name="geometry">
      <rect>
-      <x>160</x>
-      <y>70</y>
+      <x>10</x>
+      <y>20</y>
       <width>75</width>
       <height>23</height>
      </rect>
@@ -116,8 +116,8 @@
    <widget class="QLabel" name="warning_text_select_file">
     <property name="geometry">
      <rect>
-      <x>250</x>
-      <y>70</y>
+      <x>100</x>
+      <y>20</y>
       <width>161</width>
       <height>20</height>
      </rect>
@@ -133,21 +133,11 @@
      <string/>
     </property>
    </widget>
-   <widget class="QGraphicsView" name="graphicsView">
-    <property name="geometry">
-     <rect>
-      <x>40</x>
-      <y>340</y>
-      <width>256</width>
-      <height>192</height>
-     </rect>
-    </property>
-   </widget>
    <widget class="QPushButton" name="add_point">
     <property name="geometry">
      <rect>
-      <x>660</x>
-      <y>160</y>
+      <x>670</x>
+      <y>110</y>
       <width>75</width>
       <height>23</height>
      </rect>
@@ -162,8 +152,8 @@
     </property>
     <property name="geometry">
      <rect>
-      <x>480</x>
-      <y>210</y>
+      <x>490</x>
+      <y>160</y>
       <width>160</width>
       <height>22</height>
      </rect>
@@ -193,8 +183,8 @@
    <widget class="QSpinBox" name="spinBox_black_point">
     <property name="geometry">
      <rect>
-      <x>520</x>
-      <y>110</y>
+      <x>530</x>
+      <y>60</y>
       <width>81</width>
       <height>22</height>
      </rect>
@@ -212,10 +202,10 @@
    <widget class="QGroupBox" name="groupBox">
     <property name="geometry">
      <rect>
-      <x>459</x>
-      <y>69</y>
+      <x>469</x>
+      <y>19</y>
       <width>291</width>
-      <height>411</height>
+      <height>341</height>
      </rect>
     </property>
     <property name="styleSheet">
@@ -359,18 +349,44 @@
       <number>10000</number>
      </property>
     </widget>
+    <widget class="QPushButton" name="update_graph">
+     <property name="geometry">
+      <rect>
+       <x>20</x>
+       <y>300</y>
+       <width>81</width>
+       <height>23</height>
+      </rect>
+     </property>
+     <property name="text">
+      <string>Update graph</string>
+     </property>
+    </widget>
    </widget>
    <widget class="QLabel" name="Label_loader">
     <property name="geometry">
      <rect>
-      <x>170</x>
-      <y>200</y>
-      <width>47</width>
-      <height>13</height>
+      <x>30</x>
+      <y>110</y>
+      <width>81</width>
+      <height>71</height>
      </rect>
     </property>
     <property name="text">
      <string/>
+    </property>
+   </widget>
+   <widget class="QWidget" name="content_plot" native="true">
+    <property name="geometry">
+     <rect>
+      <x>30</x>
+      <y>380</y>
+      <width>731</width>
+      <height>281</height>
+     </rect>
+    </property>
+    <property name="styleSheet">
+     <string notr="true"/>
     </property>
    </widget>
    <zorder>groupBox</zorder>
@@ -380,11 +396,11 @@
    <zorder>label_4</zorder>
    <zorder>select_file</zorder>
    <zorder>warning_text_select_file</zorder>
-   <zorder>graphicsView</zorder>
    <zorder>add_point</zorder>
    <zorder>point1Slider</zorder>
    <zorder>spinBox_black_point</zorder>
    <zorder>Label_loader</zorder>
+   <zorder>content_plot</zorder>
   </widget>
   <widget class="QStatusBar" name="statusbar"/>
  </widget>
